update, deleted ready

+++ backend/src/controller/profile.controller.ts
@@ -1,57 +1,60 @@
-import { Request, Response } from "express";
+import { NextFunction, Request, Response } from "express";
 import { User } from "../models/user.model";
 import profileService from "../services/profile.service";
+import { BadRequestError } from "../errors/bad-request.error";
 
 class ProfileController {
 
-    async get  ( req : Request, res : Response ) {
+    async get  ( req : Request, res : Response, next : NextFunction ) {
         try{
             const id = req.payload?.subjectId;
             const user = await profileService.get( id! );
     
             if ( !user || !id )
-                throw new Error(`Usuario con id ${ id } no encontrado`);
+                throw new BadRequestError(`Usuario con id ${ id } no encontrado`);
 
             res.status(200).json( user );
         } catch ( error ) {
-            if ( error instanceof Error ){
-                res.status(400).json({ msg : `${ error.message }`});
-            };
+            next( error )
         }
     };
 
-    async update ( req : Request, res : Response ) {
+    async update ( req : Request, res : Response, next : NextFunction ) {
         try {
             const id = req.payload?.subjectId;
-            const updateResult = await profileService.update(id! , req.body);
 
-            if ( !id || !updateResult ) 
-                throw new Error(`Usuario con id ${ id } no existe`);
+            const user = await profileService.get( id! );
 
-            res.status(200).json( updateResult );
+            if( !user ) 
+                throw new BadRequestError(`Usuario con id ${ id } no encontrado`);
+
+            await profileService.update( id!, req.body)
+
+            const updatedUser = await profileService.get( id! );
+
+            res.status(200).json( updatedUser );
 
         } catch ( error ) {
-            if ( error instanceof Error ) {
-                res.status(400).json({ msg : `${ error.message }`});
-            };
+            next( error )
         }
 
     };
 
-    async delete ( req : Request, res : Response ) {
+    async delete ( req : Request, res : Response, next : NextFunction ) {
         try {
             const id = req.payload?.subjectId;
-            const deleteResult = await profileService.delete( id! );
-    
-            if ( !id || !deleteResult )
-                throw new Error(`Usuario con id ${ id } no existe`);
-    
-            res.status(200).json( deleteResult );
+
+            const user = await profileService.get( id! );
+
+            if( !user ) 
+                throw new BadRequestError(`Usuario con id ${ id } no encontrado`);
+
+            await profileService.delete( id! );
+
+            res.status(200).json( true );
             
         } catch ( error ) {
-            if ( error instanceof Error ) {
-                res.status(400).json({ msg : `${ error.message }`});
-            };
+            next( error )
         }
     };
